update test names, add persisted datastore value for slga migration
