+++ drivers/SmartThings/zigbee-lock/src/legacy-handlers/can_handle.lua
@@ -2,10 +2,9 @@
 -- Licensed under the Apache License, Version 2.0
 
 return function(opts, driver, device, ...)
-  local capabilities = require "st.capabilities"
-  local lock_codes_migrated = device:get_latest_state("main", capabilities.lockCodes.ID,
-    capabilities.lockCodes.migrated.NAME, false)
-  if not lock_codes_migrated then
+  local lock_utils = require "zigbee_lock_utils"
+  local slga_migrated = device:get_field(lock_utils.SLGA_MIGRATED) or false
+  if not slga_migrated then
     local subdriver = require("legacy-handlers")
     return true, subdriver
   end
