+++ drivers/SmartThings/zigbee-lock/src/zigbee_lock_utils.lua
@@ -5,7 +5,7 @@ local capabilities = require "st.capabilities"
 local utils = require "st.utils"
 local INITIAL_INDEX = 1
 
-local new_lock_utils = {
+local lock_utils = {
   -- Constants
   ADD_CREDENTIAL = "addCredential",
   ADD_USER = "addUser",
@@ -32,31 +32,32 @@ local new_lock_utils = {
   UPDATE_USER = "updateUser",
   USER_INDEX = "userIndex",
   USER_NAME = "userName",
-  USER_TYPE = "userType"
+  USER_TYPE = "userType",
+  SLGA_MIGRATED = "slgaMigrated"
 }
 
-new_lock_utils.DELETE_ALL_USERS = 0xFF
+lock_utils.DELETE_ALL_USERS = 0xFF
 
 -- check if we are currently busy performing a task.
 -- if we aren't then set as busy.
-new_lock_utils.busy_check_and_set = function (device, command, override_busy_check)
+lock_utils.busy_check_and_set = function (device, command, override_busy_check)
   if override_busy_check then
     -- the function was called by an injected command.
     return false
   end
   local c_time = os.time()
-  local busy_state = device:get_field(new_lock_utils.BUSY) or false
+  local busy_state = device:get_field(lock_utils.BUSY) or false
 
   if busy_state == false or c_time - busy_state > 10 then
-    device:set_field(new_lock_utils.COMMAND_NAME, command)
-    device:set_field(new_lock_utils.BUSY, c_time)
+    device:set_field(lock_utils.COMMAND_NAME, command)
+    device:set_field(lock_utils.BUSY, c_time)
     return false
   else
     local command_result_info = {
       commandName = command.name,
-      statusCode = new_lock_utils.STATUS_BUSY
+      statusCode = lock_utils.STATUS_BUSY
     }
-    if command.type == new_lock_utils.LOCK_USERS then
+    if command.type == lock_utils.LOCK_USERS then
       device:emit_event(capabilities.lockUsers.commandResult(
         command_result_info, { state_change = true, visibility = { displayed = true } }
       ))
@@ -69,18 +70,18 @@ new_lock_utils.busy_check_and_set = function (device, command, override_busy_che
   end
 end
 
-new_lock_utils.clear_busy_state = function(device, status, override_busy_check)
+lock_utils.clear_busy_state = function(device, status, override_busy_check)
   if override_busy_check then
     return
   end
-  local command = device:get_field(new_lock_utils.COMMAND_NAME)
-  local active_credential = device:get_field(new_lock_utils.ACTIVE_CREDENTIAL)
+  local command = device:get_field(lock_utils.COMMAND_NAME)
+  local active_credential = device:get_field(lock_utils.ACTIVE_CREDENTIAL)
   if command ~= nil then
     local command_result_info = {
       commandName = command.name,
       statusCode = status
     }
-    if command.type == new_lock_utils.LOCK_USERS then
+    if command.type == lock_utils.LOCK_USERS then
       if active_credential ~= nil and active_credential.userIndex ~= nil then
         command_result_info.userIndex = active_credential.userIndex
       end
@@ -100,35 +101,35 @@ new_lock_utils.clear_busy_state = function(device, status, override_busy_check)
     end
   end
 
-  device:set_field(new_lock_utils.ACTIVE_CREDENTIAL, nil)
-  device:set_field(new_lock_utils.COMMAND_NAME, nil)
-  device:set_field(new_lock_utils.BUSY, false)
+  device:set_field(lock_utils.ACTIVE_CREDENTIAL, nil)
+  device:set_field(lock_utils.COMMAND_NAME, nil)
+  device:set_field(lock_utils.BUSY, false)
 end
 
-new_lock_utils.reload_tables = function(device)
+lock_utils.reload_tables = function(device)
   local users = device:get_latest_state("main", capabilities.lockUsers.ID, capabilities.lockUsers.users.NAME, {})
   local credentials = device:get_latest_state("main", capabilities.lockCredentials.ID, capabilities.lockCredentials.credentials.NAME, {})
 
   if next(users) ~= nil then
-    device:set_field(new_lock_utils.LOCK_USERS, users)
+    device:set_field(lock_utils.LOCK_USERS, users)
   end
   if next(credentials) ~= nil then
-    device:set_field(new_lock_utils.LOCK_CREDENTIALS, credentials)
+    device:set_field(lock_utils.LOCK_CREDENTIALS, credentials)
   end
-  device:set_field(new_lock_utils.TABLES_LOADED, true)
+  device:set_field(lock_utils.TABLES_LOADED, true)
 end
 
-new_lock_utils.get_users = function(device)
-  if not device:get_field(new_lock_utils.TABLES_LOADED) then
-    new_lock_utils.reload_tables(device)
+lock_utils.get_users = function(device)
+  if not device:get_field(lock_utils.TABLES_LOADED) then
+    lock_utils.reload_tables(device)
   end
 
-  local users = utils.deep_copy(device:get_field(new_lock_utils.LOCK_USERS))
+  local users = utils.deep_copy(device:get_field(lock_utils.LOCK_USERS))
   return users ~= nil and users or {}
 end
 
-new_lock_utils.get_user = function(device, user_index)
-  for _, user in pairs(new_lock_utils.get_users(device)) do
+lock_utils.get_user = function(device, user_index)
+  for _, user in pairs(lock_utils.get_users(device)) do
     if user.userIndex == user_index then
       return user
     end
@@ -137,10 +138,10 @@ new_lock_utils.get_user = function(device, user_index)
   return nil
 end
 
-new_lock_utils.get_available_user_index = function(device)
+lock_utils.get_available_user_index = function(device)
   local max = device:get_latest_state("main", capabilities.lockUsers.ID,
     capabilities.lockUsers.totalUsersSupported.NAME, 0)
-  local current_users = new_lock_utils.get_users(device)
+  local current_users = lock_utils.get_users(device)
   local available_index = nil
   local used_index = {}
   for _, user in pairs(current_users) do
@@ -159,17 +160,17 @@ new_lock_utils.get_available_user_index = function(device)
   return available_index
 end
 
-new_lock_utils.get_credentials = function(device)
-  if not device:get_field(new_lock_utils.TABLES_LOADED) then
-    new_lock_utils.reload_tables(device)
+lock_utils.get_credentials = function(device)
+  if not device:get_field(lock_utils.TABLES_LOADED) then
+    lock_utils.reload_tables(device)
   end
 
-  local credentials = utils.deep_copy(device:get_field(new_lock_utils.LOCK_CREDENTIALS))
+  local credentials = utils.deep_copy(device:get_field(lock_utils.LOCK_CREDENTIALS))
   return credentials ~= nil and credentials or {}
 end
 
-new_lock_utils.get_credential = function(device, credential_index)
-  for _, credential in pairs(new_lock_utils.get_credentials(device)) do
+lock_utils.get_credential = function(device, credential_index)
+  for _, credential in pairs(lock_utils.get_credentials(device)) do
     if credential.credentialIndex == credential_index then
       return credential
     end
@@ -177,8 +178,8 @@ new_lock_utils.get_credential = function(device, credential_index)
   return nil
 end
 
-new_lock_utils.get_credential_by_user_index = function(device, user_index)
-  for _, credential in pairs(new_lock_utils.get_credentials(device)) do
+lock_utils.get_credential_by_user_index = function(device, user_index)
+  for _, credential in pairs(lock_utils.get_credentials(device)) do
     if credential.userIndex == user_index then
       return credential
     end
@@ -187,10 +188,10 @@ new_lock_utils.get_credential_by_user_index = function(device, user_index)
   return nil
 end
 
-new_lock_utils.get_available_credential_index = function(device)
+lock_utils.get_available_credential_index = function(device)
   local max = device:get_latest_state("main", capabilities.lockCredentials.ID,
     capabilities.lockCredentials.pinUsersSupported.NAME, 0)
-  local current_credentials = new_lock_utils.get_credentials(device)
+  local current_credentials = lock_utils.get_credentials(device)
   local available_index = nil
   local used_index = {}
   for _, credential in pairs(current_credentials) do
@@ -209,53 +210,53 @@ new_lock_utils.get_available_credential_index = function(device)
   return available_index
 end
 
-new_lock_utils.create_user = function(device, user_name, user_type, user_index)
+lock_utils.create_user = function(device, user_name, user_type, user_index)
   if user_name == nil then
     user_name = "Guest" .. user_index
   end
 
-  local current_users = new_lock_utils.get_users(device)
+  local current_users = lock_utils.get_users(device)
   table.insert(current_users, { userIndex = user_index, userType = user_type, userName = user_name })
-  device:set_field(new_lock_utils.LOCK_USERS, current_users, { persist = true })
+  device:set_field(lock_utils.LOCK_USERS, current_users, { persist = true })
 end
 
-new_lock_utils.delete_user = function(device, user_index)
-  local current_users = new_lock_utils.get_users(device)
-  local status_code = new_lock_utils.STATUS_FAILURE
+lock_utils.delete_user = function(device, user_index)
+  local current_users = lock_utils.get_users(device)
+  local status_code = lock_utils.STATUS_FAILURE
 
   for index, user in pairs(current_users) do
     if user.userIndex == user_index then
       -- table.remove causes issues if we are removing while iterating.
       -- instead set the value as nil and let `prep_table` handle removing it.
       current_users[index] = nil
-      device:set_field(new_lock_utils.LOCK_USERS, current_users)
-      status_code = new_lock_utils.STATUS_SUCCESS
+      device:set_field(lock_utils.LOCK_USERS, current_users)
+      status_code = lock_utils.STATUS_SUCCESS
       break
     end
   end
   return status_code
 end
 
-new_lock_utils.add_credential = function(device, user_index, credential_type, credential_index)
-  local credentials = new_lock_utils.get_credentials(device)
+lock_utils.add_credential = function(device, user_index, credential_type, credential_index)
+  local credentials = lock_utils.get_credentials(device)
   table.insert(credentials,
     { userIndex = user_index, credentialIndex = credential_index, credentialType = credential_type })
-  device:set_field(new_lock_utils.LOCK_CREDENTIALS, credentials, { persist = true })
-  return new_lock_utils.STATUS_SUCCESS
+  device:set_field(lock_utils.LOCK_CREDENTIALS, credentials, { persist = true })
+  return lock_utils.STATUS_SUCCESS
 end
 
-new_lock_utils.delete_credential = function(device, credential_index)
-  local credentials = new_lock_utils.get_credentials(device)
-  local status_code = new_lock_utils.STATUS_FAILURE
+lock_utils.delete_credential = function(device, credential_index)
+  local credentials = lock_utils.get_credentials(device)
+  local status_code = lock_utils.STATUS_FAILURE
 
   for index, credential in pairs(credentials) do
     if credential.credentialIndex == credential_index then
-      new_lock_utils.delete_user(device, credential.userIndex)
+      lock_utils.delete_user(device, credential.userIndex)
       -- table.remove causes issues if we are removing while iterating.
       -- instead set the value as nil and let `prep_table` handle removing it.
       credentials[index] = nil
-      device:set_field(new_lock_utils.LOCK_CREDENTIALS, credentials)
-      status_code = new_lock_utils.STATUS_SUCCESS
+      device:set_field(lock_utils.LOCK_CREDENTIALS, credentials)
+      status_code = lock_utils.STATUS_SUCCESS
       break
     end
   end
@@ -263,16 +264,16 @@ new_lock_utils.delete_credential = function(device, credential_index)
   return status_code
 end
 
-new_lock_utils.update_credential = function(device, credential_index, user_index, credential_type)
-  local credentials = new_lock_utils.get_credentials(device)
-  local status_code = new_lock_utils.STATUS_FAILURE
+lock_utils.update_credential = function(device, credential_index, user_index, credential_type)
+  local credentials = lock_utils.get_credentials(device)
+  local status_code = lock_utils.STATUS_FAILURE
 
   for _, credential in pairs(credentials) do
     if credential.credentialIndex == credential_index then
       credential.credentialType = credential_type
       credential.userIndex = user_index
-      device:set_field(new_lock_utils.LOCK_CREDENTIALS, credentials, { persist = true })
-      status_code = new_lock_utils.STATUS_SUCCESS
+      device:set_field(lock_utils.LOCK_CREDENTIALS, credentials, { persist = true })
+      status_code = lock_utils.STATUS_SUCCESS
       break
     end
   end
@@ -280,7 +281,7 @@ new_lock_utils.update_credential = function(device, credential_index, user_index
 end
 
 -- emit_event doesn't like having `nil` values in the table. Remove any if they are present.
-new_lock_utils.prep_table = function(data)
+lock_utils.prep_table = function(data)
   local clean_table = {}
   for _, value in pairs(data) do
     if value ~= nil then
@@ -290,19 +291,19 @@ new_lock_utils.prep_table = function(data)
   return clean_table
 end
 
-new_lock_utils.send_events = function(device, type)
-  if type == nil or type == new_lock_utils.LOCK_USERS then
-    local current_users = new_lock_utils.prep_table(new_lock_utils.get_users(device))
-    device:set_field(new_lock_utils.LOCK_USERS, current_users, { persist = true })
+lock_utils.send_events = function(device, type)
+  if type == nil or type == lock_utils.LOCK_USERS then
+    local current_users = lock_utils.prep_table(lock_utils.get_users(device))
+    device:set_field(lock_utils.LOCK_USERS, current_users, { persist = true })
     device:emit_event(capabilities.lockUsers.users(current_users,
       {state_change = true, visibility = { displayed = true } }))
   end
-  if type == nil or type == new_lock_utils.LOCK_CREDENTIALS then
-    local credentials = new_lock_utils.prep_table(new_lock_utils.get_credentials(device))
-    device:set_field(new_lock_utils.LOCK_CREDENTIALS, credentials, { persist = true })
+  if type == nil or type == lock_utils.LOCK_CREDENTIALS then
+    local credentials = lock_utils.prep_table(lock_utils.get_credentials(device))
+    device:set_field(lock_utils.LOCK_CREDENTIALS, credentials, { persist = true })
     device:emit_event(capabilities.lockCredentials.credentials(credentials,
       { state_change = true,  visibility = { displayed = true } }))
   end
 end
 
-return new_lock_utils
+return lock_utils
